requirements.txt
fastapi
uvicorn
psycopg2-binary
python-dotenv
tweepy
requests
nltk
textblob
scikit-learn
pandas